添加作业管理功能并重构活动相关模块
- 移除ActivityMapper中的注解SQL，改为XML文件实现
- 添加HomeWork实体类、DTO、VO及相关映射器和控制器
- 在ActivityVo中增加作业统计字段homeworkSum
- 修改考试标题DTO中的分数字段类型为Integer
- 优化试题选择逻辑，从按数量改为按分数抽取
- 添加作业提交的时间限制检查功能
- 实现学生作业的增删改查接口
- 优化试卷自动生成算法，使用贪心策略按分数选题
- 添加学生答案统计分析功能
- 限制只有学生角色可以设置学习进度

+++ src/main/java/com/onik/eduspring/entity/PaperQuestion.java
@@ -16,7 +16,7 @@ public class PaperQuestion {
     private Long id;
     private Long paperId;
     private Long questionId;
-    private Double score;
+    private int score;
     private Integer sort;
 
 }
